refact: 게시글 상세 모달 디자인 변경
